% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{draw_mis_vector}
\alias{draw_mis_vector}
\title{function to draw a vector from origin to misclassified point}
\usage{
draw_mis_vector(pt)
}
\arguments{
\item{pt}{point(x, y)}
}
\description{
function to draw a vector from origin to misclassified point
}
