% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{draw_mis_vector_added}
\alias{draw_mis_vector_added}
\title{function to draw weight vector + yi*xi}
\usage{
draw_mis_vector_added(w, pt, y)
}
\arguments{
\item{w}{weight vector (w0, w1, w2)}

\item{pt}{point(x, y)}

\item{y}{class of misclassified point (-1 or 1)}
}
\description{
function to draw weight vector + yi*xi
}
