% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{draw_new_weight_vector}
\alias{draw_new_weight_vector}
\title{function to draw new weight vector}
\usage{
draw_new_weight_vector(w, pt, y)
}
\arguments{
\item{w}{weight vector (w0, w1, w2)}

\item{pt}{point(x, y)}
}
\description{
function to draw new weight vector
}
