% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{draw_points}
\alias{draw_points}
\title{function to draw points}
\usage{
draw_points(X1, X2, Y, axes = FALSE)
}
\arguments{
\item{X1}{vector of values}

\item{X2}{vector of values}

\item{Y}{vector of values (-1 or 1)}

\item{axes}{logical}
}
\description{
function to draw points
}
