% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{draw_weight_vector}
\alias{draw_weight_vector}
\title{function to draw a weight vector}
\usage{
draw_weight_vector(w)
}
\arguments{
\item{w}{weight vector (w0, w1, w2)}
}
\description{
function to draw a weight vector
}
