% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_endpoints}
\alias{get_endpoints}
\title{function to get endpoints of a line as it crosses through a plot window}
\usage{
get_endpoints(W)
}
\arguments{
\item{W}{weight vector}
}
\description{
function to get endpoints of a line as it crosses through a plot window
}
