% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{intercept}
\alias{intercept}
\title{function to determine the intercept of a hyperplane}
\usage{
intercept(w)
}
\arguments{
\item{w}{weight vector (w0, w1, w2)}
}
\value{
intercept
}
\description{
function to determine the intercept of a hyperplane
}
