% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{label_vector}
\alias{label_vector}
\title{function to label a vector}
\usage{
label_vector(x0, y0, x1, y1, label)
}
\arguments{
\item{x0, y0}{coordinates of points from which to draw}

\item{x1, y1}{coordinates of points to which to draw}

\item{label}{}
}
\description{
function to label a vector
}
