% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_perceptron.R
\name{predict_perceptron}
\alias{predict_perceptron}
\title{Predict using Perceptron Model}
\usage{
predict_perceptron(model, X, threshold = 0)
}
\arguments{
\item{model}{Trained perceptron model}

\item{X}{Input numeric matrix of features}

\item{threshold}{Threshold value for prediction (default: 0)}
}
\value{
Predicted labels
}
\description{
This function makes predictions using the trained perceptron model.
}
\examples{
predict_perceptron(model, data$X)
}
