% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shift}
\alias{shift}
\title{function to determine the intersection ("shift") of a weight vector and hyperplane}
\usage{
shift(w)
}
\arguments{
\item{w}{weight vector (w0, w1, w2)}
}
\value{
\code{shift()} returns a named vector: "x" and "y"
}
\description{
function to determine the intersection ("shift") of a weight vector and hyperplane
}
