% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{slope}
\alias{slope}
\title{function to determine the slope of a hyperplane}
\usage{
slope(w)
}
\arguments{
\item{w}{weight vector (w0, w1, w2)}
}
\value{
slope
}
\description{
function to determine the slope of a hyperplane
}
