% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_perceptron.R
\name{train_perceptron}
\alias{train_perceptron}
\title{Perceptron Algorithm}
\usage{
train_perceptron(
  X,
  y,
  max_iter = 100,
  eta = 0.01,
  acc_th = 0.95,
  threshold = 0
)
}
\arguments{
\item{X}{Input numeric matrix of features. Where each row represents a sample and each column represents a feature.}

\item{y}{Target binary numeric vector. Represents the binary target labels (0 or 1) corresponding to the samples in X.}

\item{max_iter}{Maximum number of epochs (iterations) for training the perceptron model.}

\item{eta}{Learning rate (default: 0.01).}

\item{acc_th}{Threshold for accuracy to break the training loop (default: 0.95).}

\item{threshold}{Threshold value for prediction (default: 0).}
}
\value{
Trained perceptron model.
}
\description{
This function implements the perceptron algorithm.
}
\examples{
data(synthetic_dataset)
train_perceptron(synthetic_dataset[, c("X1", "X2")], synthetic_dataset$y)
}
