% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_decision_boundary.R
\name{visualize_decision_boundary}
\alias{visualize_decision_boundary}
\title{Visualize Decision Boundary}
\usage{
visualize_decision_boundary(model, X, y)
}
\arguments{
\item{model}{Trained perceptron model}

\item{X}{Input numeric matrix of features}

\item{y}{Target vector}
}
\value{
ggplot object
}
\description{
This function visualizes the decision boundary of the perceptron model, which separates the classes based on the learned weights.
The decision boundary represents the line (or hyperplane in higher dimensions) where the perceptron model's output changes from one class to another.
}
\examples{
visualize_decision_boundary(model, data$X, data$y)
}
